Update: button component
