+++ src/components/index.ts
@@ -4,3 +4,5 @@ export * from './MobileMenuToggle';
 export * from './MobileMenu';
 export * from './LanguageSelector';
 export * from './BackgroundElement';
+export * from './Button';
+export * from './ProjectCard';
