+++ src/sections/Wellcome/index.tsx
@@ -1,5 +1,5 @@
 import { BackgroundElement } from "../../components";
-import Button from "../../components/Button/Button";
+import { Button } from "../../components/";
 
 export const Wellcome = () => {
   return (
